requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
schedule
email-validator
exa_py
sumy
beautifulsoup4
lxml
numpy